completed react-multiple-components
