+++ react-multiple-components/my-app/src/Banner.tsx
@@ -1,0 +1,7 @@
+type bannerProps = {
+  item: string;
+};
+
+export function Banner({ item }: bannerProps) {
+  return <h2 id="banner">{item}</h2>;
+}
