+++ react-multiple-components/my-app/src/Indicators.tsx
@@ -1,37 +1,3 @@
-import React from 'react';
-
-type BannerProps = {
-  item: string;
-};
-
-export function Banner({ item }: BannerProps) {
-  return <h2 id="banner">{item}</h2>;
-}
-
-type nextButtonProps = {
-  onNext: () => void;
-};
-
-export function NextButton({ onNext }: nextButtonProps) {
-  return (
-    <button id="next" onClick={onNext}>
-      Next
-    </button>
-  );
-}
-
-type previousButtonProps = {
-  onPrevious: () => void;
-};
-
-export function PreviousButton({ onPrevious }: previousButtonProps) {
-  return (
-    <button id="previous" onClick={onPrevious}>
-      Previous
-    </button>
-  );
-}
-
 type indicatorsProps = {
   count: number;
   highlightedIndex: number;
