+++ react-multiple-components/my-app/src/NextButton.tsx
@@ -1,0 +1,11 @@
+type nextButtonProps = {
+  onNext: () => void;
+};
+
+export function NextButton({ onNext }: nextButtonProps) {
+  return (
+    <button id="next" onClick={onNext}>
+      Next
+    </button>
+  );
+}
