+++ react-multiple-components/my-app/src/PreviousButton.tsx
@@ -1,0 +1,11 @@
+type previousButtonProps = {
+  onPrevious: () => void;
+};
+
+export function PreviousButton({ onPrevious }: previousButtonProps) {
+  return (
+    <button id="previous" onClick={onPrevious}>
+      Previous
+    </button>
+  );
+}
